+++ iMEGA/Camera uploads/UploadOperations/CameraUploadOperation.m
@@ -143,7 +143,7 @@
 - (void)requestUploadURL {
     [[MEGASdkManager sharedMEGASdk] requestBackgroundUploadURLWithFileSize:self.uploadInfo.fileSize mediaUpload:self.uploadInfo.mediaUpload delegate:[[CameraUploadRequestDelegate alloc] initWithCompletion:^(MEGARequest * _Nonnull request, MEGAError * _Nonnull error) {
         if (error.type) {
-            MEGALogError(@"[Camera Upload] %@ requests upload url failed with error type: %ld", self, error.type);
+            MEGALogError(@"[Camera Upload] %@ requests upload url failed with error type: %ld", self, (long)error.type);
             [self finishOperationWithStatus:CameraAssetUploadStatusFailed shouldUploadNextAsset:YES];
         } else {
             self.uploadInfo.uploadURLString = [self.uploadInfo.mediaUpload uploadURLString];
